+++ nyx/engine/managers/nyx_entity_manager.py
